requirements.txt
aioredis==1.0.0
asgi-redis==1.0.0
asgiref==2.1.6
async-timeout==2.0.1
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
cached-property==1.5.1
channels==2.0.2
channels-redis==2.1.0
constantly==15.1.0
daphne==2.2.4
Django==2.1.5
django-rest-framework==0.1.0
django-url-filter==0.3.12
djangorestframework==3.9.1
enum-compat==0.0.2
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
msgpack==0.5.6
msgpack-python==0.5.6
PyHamcrest==1.9.0
pytz==2018.9
redis==2.10.6
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
zope.interface==4.6.0
